Better code organisation
Now we have a seperate Flite.Projections.Conversion module
that deals with converting the 'normal' Flite types and functions
into those that are suitable for projections analysis.

The plan is to write _only_ the projection analysis in Flite.Projections
and write any additional modules to facilitate that. I'm trying to have
good organisation :)